+++ test/binary.js
@@ -21,7 +21,7 @@ describe('Binary', function () {
     this.timeout(8000);
     const {stdout, stderr} = await execFile(
       binFile,
-      ['-l', 'test/fixtures/licenseInfo.json', 'test.svg']
+      ['-l', 'test/fixtures/licenseInfo.json', '--packagePath', '.', 'test.svg']
     );
     if (stderr) {
       // eslint-disable-next-line no-console